+++ ED-BSI/List-1/c-lang/stack.h
@@ -10,8 +10,8 @@ void destroy_stack(Stack *s);
 bool is_empty(const Stack *s);
 bool is_full(const Stack *s);
 void make_empty(Stack *s);
-bool push(Stack *s, int value);
-int peek(Stack *s);
-int pop(Stack *s);
+bool push(Stack *s, long int value);
+long int peek(Stack *s);
+long int pop(Stack *s);
 
 #endif
